GemChat/requirements.txt
google-generativeai 
python-dotenv 
pdfplumber 
pandas
flask